// sdks/node-ts/src/apache_beam/proto/beam_fn_api.ts
export interface RemoteGrpcPort {
    /**
     * (Required) An API descriptor which describes where to
     * connect to including any authentication that is required.
     *
     * @generated from protobuf field: org.apache.beam.model.pipeline.v1.ApiServiceDescriptor api_service_descriptor = 1;
     */
    apiServiceDescriptor?: ApiServiceDescriptor;
    /**
     * (Required) The ID of the Coder that will be used to encode and decode data
     * sent over this port.
     *
     * @generated from protobuf field: string coder_id = 2;
     */
    coderId: string;
}
// ...
export interface InstructionRequest {
    /**
     * (Required) A unique identifier provided by the runner which represents
     * this requests execution. The InstructionResponse MUST have the matching id.
     *
     * @generated from protobuf field: string instruction_id = 1;
     */
    instructionId: string;
    /**
     * @generated from protobuf oneof: request
     */
    request: {
        oneofKind: "processBundle";
        /**
         * @generated from protobuf field: org.apache.beam.model.fn_execution.v1.ProcessBundleRequest process_bundle = 1001;
         */
        processBundle: ProcessBundleRequest;
    } | {
        oneofKind: "processBundleProgress";
        /**
         * @generated from protobuf field: org.apache.beam.model.fn_execution.v1.ProcessBundleProgressRequest process_bundle_progress = 1002;
         */
        processBundleProgress: ProcessBundleProgressRequest;
    } | {
        oneofKind: "processBundleSplit";
        /**
         * @generated from protobuf field: org.apache.beam.model.fn_execution.v1.ProcessBundleSplitRequest process_bundle_split = 1003;
         */
        processBundleSplit: ProcessBundleSplitRequest;
    } | {
        oneofKind: "finalizeBundle";
        /**
         * @generated from protobuf field: org.apache.beam.model.fn_execution.v1.FinalizeBundleRequest finalize_bundle = 1004;
         */
        finalizeBundle: FinalizeBundleRequest;
    } | {
        oneofKind: "monitoringInfos";
        /**
         * @generated from protobuf field: org.apache.beam.model.fn_execution.v1.MonitoringInfosMetadataRequest monitoring_infos = 1005;
         */
        monitoringInfos: MonitoringInfosMetadataRequest;
    } | {
        oneofKind: "harnessMonitoringInfos";
        /**
         * @generated from protobuf field: org.apache.beam.model.fn_execution.v1.HarnessMonitoringInfosRequest harness_monitoring_infos = 1006;
         */
        harnessMonitoringInfos: HarnessMonitoringInfosRequest;
    } | {
        oneofKind: "register";
        /**
         * DEPRECATED
         *
         * @generated from protobuf field: org.apache.beam.model.fn_execution.v1.RegisterRequest register = 1000;
         */
        register: RegisterRequest;
    } | {
        oneofKind: undefined;
    };
}
/**
 * The response for an associated request the SDK had been asked to fulfill.
 * Stable
 *
 * @generated from protobuf message org.apache.beam.model.fn_execution.v1.InstructionResponse
 */
export interface InstructionResponse {
    /**
     * (Required) A reference provided by the runner which represents a requests
     * execution. The InstructionResponse MUST have the matching id when
     * responding to the runner.
     *
     * @generated from protobuf field: string instruction_id = 1;
     */
    instructionId: string;
    /**
     * If this is specified, then this instruction has failed.
     * A human readable string representing the reason as to why processing has
     * failed.
     *
     * @generated from protobuf field: string error = 2;
     */
    error: string;
    /**
     * @generated from protobuf oneof: response
     */
    response: {
        oneofKind: "processBundle";
        /**
         * @generated from protobuf field: org.apache.beam.model.fn_execution.v1.ProcessBundleResponse process_bundle = 1001;
         */
        processBundle: ProcessBundleResponse;
    } | {
        oneofKind: "processBundleProgress";
        /**
         * @generated from protobuf field: org.apache.beam.model.fn_execution.v1.ProcessBundleProgressResponse process_bundle_progress = 1002;
         */
        processBundleProgress: ProcessBundleProgressResponse;
    } | {
        oneofKind: "processBundleSplit";
        /**
         * @generated from protobuf field: org.apache.beam.model.fn_execution.v1.ProcessBundleSplitResponse process_bundle_split = 1003;
         */
        processBundleSplit: ProcessBundleSplitResponse;
    } | {
        oneofKind: "finalizeBundle";
        /**
         * @generated from protobuf field: org.apache.beam.model.fn_execution.v1.FinalizeBundleResponse finalize_bundle = 1004;
         */
        finalizeBundle: FinalizeBundleResponse;
    } | {
        oneofKind: "monitoringInfos";
        /**
         * @generated from protobuf field: org.apache.beam.model.fn_execution.v1.MonitoringInfosMetadataResponse monitoring_infos = 1005;
         */
        monitoringInfos: MonitoringInfosMetadataResponse;
    } | {
        oneofKind: "harnessMonitoringInfos";
        /**
         * @generated from protobuf field: org.apache.beam.model.fn_execution.v1.HarnessMonitoringInfosResponse harness_monitoring_infos = 1006;
         */
        harnessMonitoringInfos: HarnessMonitoringInfosResponse;
    } | {
        oneofKind: "register";
        /**
         * DEPRECATED
         *
         * @generated from protobuf field: org.apache.beam.model.fn_execution.v1.RegisterResponse register = 1000;
         */
        register: RegisterResponse;
    } | {
        oneofKind: undefined;
    };
}
// ...
export interface RegisterResponse {
}
/**
 * Definitions that should be used to construct the bundle processing graph.
 *
 * @generated from protobuf message org.apache.beam.model.fn_execution.v1.ProcessBundleDescriptor
 */
export interface ProcessBundleDescriptor {
    /**
     * (Required) A pipeline level unique id which can be used as a reference to
     * refer to this.
     *
     * @generated from protobuf field: string id = 1;
     */
    id: string;
    /**
     * (Required) A map from pipeline-scoped id to PTransform.
     *
     * @generated from protobuf field: map<string, org.apache.beam.model.pipeline.v1.PTransform> transforms = 2;
     */
    transforms: {
        [key: string]: PTransform;
    };
    /**
     * (Required) A map from pipeline-scoped id to PCollection.
     *
     * @generated from protobuf field: map<string, org.apache.beam.model.pipeline.v1.PCollection> pcollections = 3;
     */
    pcollections: {
        [key: string]: PCollection;
    };
    /**
     * (Required) A map from pipeline-scoped id to WindowingStrategy.
     *
     * @generated from protobuf field: map<string, org.apache.beam.model.pipeline.v1.WindowingStrategy> windowing_strategies = 4;
     */
    windowingStrategies: {
        [key: string]: WindowingStrategy;
    };
    /**
     * (Required) A map from pipeline-scoped id to Coder.
     *
     * @generated from protobuf field: map<string, org.apache.beam.model.pipeline.v1.Coder> coders = 5;
     */
    coders: {
        [key: string]: Coder;
    };
    /**
     * (Required) A map from pipeline-scoped id to Environment.
     *
     * @generated from protobuf field: map<string, org.apache.beam.model.pipeline.v1.Environment> environments = 6;
     */
    environments: {
        [key: string]: Environment;
    };
    /**
     * A descriptor describing the end point to use for State API
     * calls. Required if the Runner intends to send remote references over the
     * data plane or if any of the transforms rely on user state or side inputs.
     *
     * @generated from protobuf field: org.apache.beam.model.pipeline.v1.ApiServiceDescriptor state_api_service_descriptor = 7;
     */
    stateApiServiceDescriptor?: ApiServiceDescriptor;
    /**
     * A descriptor describing the end point to use for Data API for user timers.
     * Required if the ProcessBundleDescriptor contains any transforms that have user timers.
     *
     * @generated from protobuf field: org.apache.beam.model.pipeline.v1.ApiServiceDescriptor timer_api_service_descriptor = 8;
     */
    timerApiServiceDescriptor?: ApiServiceDescriptor;
}
// ...
export interface ProcessBundleRequest {
    /**
     * (Required) A reference to the process bundle descriptor that must be
     * instantiated and executed by the SDK harness.
     *
     * @generated from protobuf field: string process_bundle_descriptor_id = 1;
     */
    processBundleDescriptorId: string;
    /**
     * (Optional) A list of cache tokens that can be used by an SDK to reuse
     * cached data returned by the State API across multiple bundles.
     *
     * @generated from protobuf field: repeated org.apache.beam.model.fn_execution.v1.ProcessBundleRequest.CacheToken cache_tokens = 2;
     */
    cacheTokens: ProcessBundleRequest_CacheToken[];
    /**
     * (Optional) Elements to be processed with the bundle. Either all or
     * none of the bundle elements should be included in the ProcessBundleRequest.
     * This embedding is to achieve better efficiency for bundles that contain
     * only small amounts of data and are cheap to be processed on the SDK harness
     * side. This field can be set only if the SDK declares that it supports the
     * beam:protocol:control_request_elements_embedding:v1 capability. See more
     * at https://s.apache.org/beam-fn-api-control-data-embedding.
     *
     * @generated from protobuf field: org.apache.beam.model.fn_execution.v1.Elements elements = 3;
     */
    elements?: Elements;
}
// ...
export interface ProcessBundleProgressRequest {
    /**
     * (Required) A reference to an active process bundle request with the given
     * instruction id.
     *
     * @generated from protobuf field: string instruction_id = 1;
     */
    instructionId: string;
}
// ...
export interface FinalizeBundleRequest {
    /**
     * (Required) A reference to a completed process bundle request with the given
     * instruction id.
     *
     * @generated from protobuf field: string instruction_id = 1;
     */
    instructionId: string;
}
// ...
export interface StartWorkerRequest {
    /**
     * @generated from protobuf field: string worker_id = 1;
     */
    workerId: string;
    /**
     * @generated from protobuf field: org.apache.beam.model.pipeline.v1.ApiServiceDescriptor control_endpoint = 2;
     */
    controlEndpoint?: ApiServiceDescriptor;
    /**
     * @generated from protobuf field: org.apache.beam.model.pipeline.v1.ApiServiceDescriptor logging_endpoint = 3;
     */
    loggingEndpoint?: ApiServiceDescriptor;
    /**
     * @generated from protobuf field: org.apache.beam.model.pipeline.v1.ApiServiceDescriptor artifact_endpoint = 4;
     */
    artifactEndpoint?: ApiServiceDescriptor;
    /**
     * @generated from protobuf field: org.apache.beam.model.pipeline.v1.ApiServiceDescriptor provision_endpoint = 5;
     */
    provisionEndpoint?: ApiServiceDescriptor;
    /**
     * @generated from protobuf field: map<string, string> params = 10;
     */
    params: {
        [key: string]: string;
    };
}
/**
 * @generated from protobuf message org.apache.beam.model.fn_execution.v1.StartWorkerResponse
 */
export interface StartWorkerResponse {
    /**
     * @generated from protobuf field: string error = 1;
     */
    error: string;
}
/**
 * @generated from protobuf message org.apache.beam.model.fn_execution.v1.StopWorkerRequest
 */
export interface StopWorkerRequest {
    /**
     * @generated from protobuf field: string worker_id = 1;
     */
    workerId: string;
}
/**
 * @generated from protobuf message org.apache.beam.model.fn_execution.v1.StopWorkerResponse
 */
export interface StopWorkerResponse {
    /**
     * @generated from protobuf field: string error = 1;
     */
    error: string;
}
// ...
export const RemoteGrpcPort = new RemoteGrpcPort$Type();
// ...
export const InstructionRequest = new InstructionRequest$Type();
// ...
export const InstructionResponse = new InstructionResponse$Type();
// ...
export const ProcessBundleDescriptor = new ProcessBundleDescriptor$Type();
// ...
export const ProcessBundleRequest = new ProcessBundleRequest$Type();
// ...
export const ProcessBundleProgressRequest = new ProcessBundleProgressRequest$Type();
// ...
export const MonitoringInfosMetadataRequest = new MonitoringInfosMetadataRequest$Type();
// ...
export const ProcessBundleSplitRequest = new ProcessBundleSplitRequest$Type();
// ...
export const FinalizeBundleRequest = new FinalizeBundleRequest$Type();
// ...
export const StartWorkerRequest = new StartWorkerRequest$Type();
// ...
export const StartWorkerResponse = new StartWorkerResponse$Type();
// ...
export const StopWorkerRequest = new StopWorkerRequest$Type();
// ...
export const StopWorkerResponse = new StopWorkerResponse$Type();